pyproject.toml
[tool.poetry]
name = "amcef"
version = "1.0.0"
description = ""
authors = ["ItsDrike <itsdrike@protonmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
repository = "https://github.com/ItsDrike/AMCEF"

[tool.poetry.dependencies]
python = "^3.10"
taskipy = "^1.10.2"
fastapi = "^0.78.0"
uvicorn = "^0.18.2"
coloredlogs = "^15.0.1"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-annotations = "^2.9.0"
flake8-bugbear = "^22.7.1"
flake8-pep585 = "^0.1.5"
flake8-simplify = "^0.19.2"
flake8-tidy-imports = "^4.8.0"
pep8-naming = "^0.13.0"
black = "^22.6.0"
isort = "^5.10.1"
pyright = "^1.1.259"
pre-commit = "^2.20.0"

[tool.black]
line-length = 119
extend-exclude = "^/.cache"

[tool.taskipy.tasks]
precommit = "pre-commit run install"
lint = "pre-commit run --all-files"
black = "black ."
isort = "isort ."
pyright = "pyright ."
flake8 = "flake8 ."
run = "uvicorn src.__init__:app"
run-dev = "DEBUG=1 uvicorn --reload src.__init__:app"

[tool.isort]
profile = "black"
line_length = 119
atomic = true
order_by_type = false
case_sensitive = true
combine_as_imports = true
skip = [".venv", ".git", ".cache"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
